Move write into service directory
