+++ tests/Service/WriteTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Jstewmc\Rtf;
+namespace Jstewmc\Rtf\Service;
+
+use Jstewmc\Rtf\Element;
 
 class WriteTest extends \PHPUnit\Framework\TestCase
 {
@@ -25,7 +27,7 @@ class WriteTest extends \PHPUnit\Framework\TestCase
                 ->appendChild(new Element\Text('bar'))
                 ->appendChild(new Element\Control\Word\B(0)));
 
-        $root = (new Service\Render())($group);
+        $root = (new Render())($group);
 
         $this->assertEquals('{foo {\b bar\b0 }}', (new Write())($root));
     }
@@ -39,7 +41,7 @@ class WriteTest extends \PHPUnit\Framework\TestCase
                 ->appendChild(new Element\Text('bar'))
                 ->appendChild(new Element\Control\Word\B(0)));
 
-        $root = (new Service\Render())($group);
+        $root = (new Render())($group);
 
         $expected = '<section style=""><p style=""><span style="">foo </span>'
             . '<span style="font-weight: bold;">bar</span></p></section>';
@@ -57,7 +59,7 @@ class WriteTest extends \PHPUnit\Framework\TestCase
                 ->appendChild(new Element\Text('bar'))
                 ->appendChild(new Element\Control\Word\B(0)));
 
-        $root = (new Service\Render())($group);
+        $root = (new Render())($group);
 
         $this->assertEquals('{foo {\b bar\b0 }}', (new Write())($root, 'rtf'));
     }
@@ -71,7 +73,7 @@ class WriteTest extends \PHPUnit\Framework\TestCase
                 ->appendChild(new Element\Text('bar'))
                 ->appendChild(new Element\Control\Word\B(0)));
 
-        $root = (new Service\Render())($group);
+        $root = (new Render())($group);
 
         $this->assertEquals('foo bar', (new Write())($root, 'text'));
     }
